requirements/dev.txt
# SHA1:72dc9f7610468db23616270e931e4130c4b0f134
#
# This file is autogenerated by pip-compile-multi
# To update, run:
#
#    pip-compile-multi
#
-r prod.txt
astroid==2.0.4
bandit==1.6.2
coverage==4.5.4           # via coveralls
coveralls==1.8.2
docopt==0.6.2             # via coveralls
entrypoints==0.3          # via flake8
factory-boy==2.12.0
faker==2.0.3              # via factory-boy
flake8==3.7.8
freezegun==0.3.12
gitdb2==2.0.6             # via gitpython
gitpython==3.0.3          # via bandit
isort==4.3.21             # via pylint
lazy-object-proxy==1.4.2  # via astroid
mccabe==0.6.1             # via flake8, pylint
multidict==4.5.2          # via yarl
mypy-extensions==0.4.2    # via mypy
mypy==0.740
parameterized==0.7.0
pbr==5.4.3                # via stevedore
pydot==1.4.1
pyflakes==2.1.1           # via flake8
pylint-django==2.0.11
pylint-plugin-utils==0.6  # via pylint-django
pylint==2.1.1
pyparsing==2.4.2          # via pydot
pyyaml==5.1.2
recordclass==0.12.0.1
smmap2==2.0.5             # via gitdb2
stevedore==1.31.0         # via bandit
text-unidecode==1.3       # via faker
typed-ast==1.4.0          # via mypy
typing-extensions==3.7.4  # via mypy
vcrpy-unittest==0.1.7
vcrpy==2.1.0              # via vcrpy-unittest
wrapt==1.11.2             # via astroid, vcrpy
yarl==1.3.0               # via vcrpy
